pyproject.toml
[project]
name = "parsimony-core"
version = "0.1.0a1"
description = "Connector framework for financial data — search, fetch, and catalog series from FRED, SDMX, FMP, SEC Edgar, and more"
authors = [{ name = "Ockham.sh", email = "team@ockham.sh" }]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["finance", "data", "connectors", "fred", "sdmx", "catalog"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.11.1,<3",
    "pandas>=2.3.0,<3",
    "pyarrow>=23.0.1",
    "httpx>=0.28.1",
    "platformdirs>=4.0.0,<5",
    # Catalog bundle runtime (HF Parquet + FAISS distribution).
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "huggingface_hub>=0.25.0",
]

[project.optional-dependencies]
# Legacy SQLite-vec + hosted embedding API path. Retained as an extra for
# users who want local SQLite-based indexing instead of HF bundles.
search = ["litellm>=1.59.0,<2", "sqlite-vec>=0.1.6"]
sec = ["edgartools>=4.0.0"]
financial-reports = ["financial-reports-generated-client>=0.1.0"]
mcp = ["mcp>=1.0.0", "tabulate>=0.9.0"]
all = ["parsimony-core[search,sec,financial-reports,mcp]"]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=5.0",
    "ruff>=0.15.10",
    "mypy>=1.10",
    "types-requests>=2.31",
    "pip-audit>=2.7",
    # Extracted plugins — included in dev so in-repo integration tests still
    # exercise the full connector surface after Phase 2+ extractions.
    "parsimony-fred",
]
docs = ["mkdocs-material"]

[project.urls]
Homepage = "https://parsimony.dev"
Repository = "https://github.com/ockham-sh/parsimony"
Documentation = "https://docs.parsimony.dev"

[project.scripts]
parsimony = "parsimony.cli:main"

[tool.uv.sources]
# Resolve extracted plugins from sibling checkouts during local dev. CI and
# published installs use the PyPI versions (pinned in [project.optional-dependencies].dev).
parsimony-fred = { path = "../parsimony-fred", editable = true }

# ---------------------------------------------------------------------------
# Bundled providers registered as entry points.
#
# Every connector currently shipped in this package is exposed via the
# ``parsimony.providers`` entry-point group so that ``parsimony.plugins.discovery``
# picks it up the same way it picks up external ``parsimony-<name>`` plugins.
#
# During the Phase 5 bulk extraction each entry listed below will be deleted
# from this file as the corresponding connector moves to its own repo
# (see ``PLAN-plugin-migration.md``).
# ---------------------------------------------------------------------------
[project.entry-points."parsimony.providers"]
# fred — extracted to parsimony-fred (PyPI / ockham-sh/parsimony-fred) in Phase 2
fmp = "parsimony.connectors.fmp"
fmp_screener = "parsimony.connectors.fmp_screener"
eodhd = "parsimony.connectors.eodhd"
coingecko = "parsimony.connectors.coingecko"
finnhub = "parsimony.connectors.finnhub"
tiingo = "parsimony.connectors.tiingo"
financial_reports = "parsimony.connectors.financial_reports"
eia = "parsimony.connectors.eia"
bdf = "parsimony.connectors.bdf"
alpha_vantage = "parsimony.connectors.alpha_vantage"
riksbank = "parsimony.connectors.riksbank"
destatis = "parsimony.connectors.destatis"
bls = "parsimony.connectors.bls"
# sdmx — extracted to parsimony-sdmx (PyPI / ockham-sh/parsimony-sdmx).
polymarket = "parsimony.connectors.polymarket"
sec_edgar = "parsimony.connectors.sec_edgar"
treasury = "parsimony.connectors.treasury"
snb = "parsimony.connectors.snb"
rba = "parsimony.connectors.rba"
bde = "parsimony.connectors.bde"
boc = "parsimony.connectors.boc"
boj = "parsimony.connectors.boj"
bdp = "parsimony.connectors.bdp"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "financial_reports_generated_client.*"
follow_imports = "skip"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["parsimony"]

[tool.hatch.build.targets.sdist]
include = ["parsimony", "README.md", "LICENSE", "CHANGELOG.md"]

[tool.pytest.ini_options]
addopts = "--import-mode=importlib --cov=parsimony --cov-report=term-missing --cov-fail-under=80"
asyncio_mode = "auto"
pythonpath = ["."]
markers = [
    "integration: hits live APIs (may be slow, requires env vars)",
]

[tool.coverage.run]
omit = [
    "parsimony/connectors/alpha_vantage.py",
    "parsimony/connectors/bde.py",
    "parsimony/connectors/bdf.py",
    "parsimony/connectors/bdp.py",
    "parsimony/connectors/bls.py",
    "parsimony/connectors/boc.py",
    "parsimony/connectors/boj.py",
    "parsimony/connectors/coingecko.py",
    "parsimony/connectors/destatis.py",
    "parsimony/connectors/eia.py",
    "parsimony/connectors/eodhd.py",
    "parsimony/connectors/financial_reports.py",
    "parsimony/connectors/finnhub.py",
    "parsimony/connectors/fmp.py",
    "parsimony/connectors/fmp_screener.py",
    "parsimony/connectors/polymarket.py",
    "parsimony/connectors/rba.py",
    "parsimony/connectors/riksbank.py",
    "parsimony/connectors/sec_edgar.py",
    "parsimony/connectors/snb.py",
    "parsimony/connectors/tiingo.py",
    "parsimony/connectors/treasury.py",
    "parsimony/mcp/__main__.py",
    "parsimony/embeddings/__init__.py",
    # Requires a real HuggingFace-hosted model (torch + sentence-transformers)
    # to exercise; tests use FakeEmbeddingProvider (tests/hf_bundle_helpers.py).
    # Loading logic is covered by the live publish pipeline, not unit tests.
    "parsimony/embeddings/sentence_transformers.py",
    # Litellm path is an optional extra — integration-tested elsewhere.
    "parsimony/embeddings/litellm.py",
]
